requirements.txt
# ----------------------------
# Core
# ----------------------------
python-dotenv==1.0.1
PyYAML==6.0.1

# ----------------------------
# Data & NLP
# ----------------------------
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.9.1
emoji==2.12.1

# ----------------------------
# Deep Learning
# ----------------------------
tensorflow==2.15.0

# ----------------------------
# API & Serving
# ----------------------------
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4

# ----------------------------
# Utilities
# ----------------------------
requests==2.32.3
